Added main menu background. Changed menu entities to blue ones. New Game button starts new game. Pressing escape gets into the main menu. WInning the game gets into main menu.

+++ courses/prog_base_3/project/consoleLevelGenerator/consoleLevelGenerator.cpp
@@ -64,7 +64,7 @@ void generateLevel(int fieldSize, float wallDensity, string path)
 	char * levelArray = new char[fieldSize*fieldSize]();
 	output.open(path, fstream::out);
 	output << to_string(fieldSize) << "\n";
-	int numberOfWalls = wallDensity * (fieldSize * fieldSize);
+	int numberOfWalls = (int)(wallDensity * (fieldSize * fieldSize));
 	for (int i = 0; i < fieldSize * fieldSize; i++)
 		levelArray[i] = NOTHING_POS_CODE;
 	for (int i = 0; i < numberOfWalls; i++)
